pyproject.toml
[project]
name = "twitch-clips"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.3.2",
    "requests>=2.32.5",
    "selenium>=4.35.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "python-dotenv>=1.1.1",
]
